+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Common and useful classes, methods, exceptions etc.
 
+# 0.1.2
+
+1. Documentation > Value Objects
+2. Docker > improve performance
+3. Utilities > Reflection > setPropertyValue() method > sets value of given property
+
 # 0.1.1
 
 1. TravisCI > run using PHP 7.2 too